Feat search more

+++ app/src/main/java/com/omoolen/omooroid/search/data/Item.kt
@@ -6,7 +6,7 @@ data class Item(
     val brand: String,
     val changeCycleMaximum: Int,
     val changeCycleMinimum: Int,
-    val diameter: Int,
+    val diameter: Double,
     val id: String,
     val imageList: List<String>,
     val name: String,
